+++ src/main/java/com/example/searchengine_ver1/backendapi/controller/SuggestionController.java
@@ -4,6 +4,9 @@ import com.example.searchengine_ver1.backendapi.service.SuggestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+import java.util.Map;
+
 @RestController
 @RequestMapping("/api/suggest")
 public class SuggestionController {
@@ -12,7 +15,8 @@ public class SuggestionController {
     private SuggestionService suggestionService;
 
     @GetMapping
-    public String getSuggestion() {
-        return suggestionService.suggest();
+    public Map<String, List<String>> getSuggestion() {
+        return suggestionService.getStructuredSuggestions();
     }
+
 }